backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
sqlite3
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
aiortc==1.5.0
python-socketio==5.10.0
eventlet==0.35.1
python-dotenv==1.0.0